+++ src/Server/server.interface.ts
@@ -1,14 +1,7 @@
-import { IncomingMessage, ServerResponse } from "http";
-
 interface IServer {
-  handleRequest: (
-    req: IncomingMessage,
-    res: ServerResponse,
-    route: Route,
-    pathVariables: Record<string, string>,
-    queryParams: Record<string, string>
-  ) => void;
-  listen: (port: number) => void;
+  handleRequest: HandleRequestFunction;
+  listen: ListenFunction;
+  response: ResponseFunction;
 }
 
 export default IServer;