+++ src/typings/global.d.ts
@@ -1,0 +1,15 @@
+type FILE_EXTENSIONS =
+  | ".html"
+  | ".js"
+  | ".css"
+  | ".json"
+  | ".png"
+  | ".jpg"
+  | ".gif"
+  | ".wav"
+  | ".mp4"
+  | ".woff"
+  | ".ttf"
+  | ".eot"
+  | ".otf"
+  | ".svg";