+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  
 ### Added
 
+- Creating Simpler server README.md file.
+
+- Creating Simpler server response method.
+
+- Creating Simpler server methods to serve static files.
+
 - Creating Simpler server base structure (Server, Router and Logger) to deal with API routes.
 
 - Adding github actions, changelog and readme files.